Add game function
A function to play the game was added, inside this function there is the
play round function that it's called 5 times using a for loop. Inside
the play round functions the getValue() functions are called to get each
play choice per round, and after this the winner of the round is
evaluated and gets it's score updated (+1).